Add Sell vehicle function to vehicle shop + add tint customization in customs shop
Was a big request for this release to have a sell function so here you guys go! I also added better controller support to the menus!
There could probably be LOTS of improvement. Im still learning lua. But it works!

Added some of my improvements to the ui as well. Many of these changes still need to be tested more.

Also added exports to player.lua to allow for use of esx scripts with pay functions.