Update DriftVehShop.lua
Add players selected/bought vehicle props to the sell menus preview vehicle.
Also added delete entity for when these submenus are closed to help increase performance. This also forces the vehicles to reload when switching between buy and sell. So that way if the same vehicle is selected in both menus it will respawn the preview vehicle with correct props.