Prevent "nil" crew creation.
Previously crew creation used "or" to check if name ~= nil or "".
This would allow users to create a blank crew name. Which would then pass it as nil.
This especially created problems when a tag was nil but name was not. Allowing the player to create the crew with a nil tag. This would cause nil errors everytime you try to call this "tag".
Using "and" we make sure we are checking both values everytime. Not just one or the other.